requirements.txt
pymongo==3.2.1
click==6.2
docker-py==1.7.0
requests==2.9.1
eventlet==0.18.3
requests==2.9.1
